[추가]마이페이지 관련 [수정] 예약xml, service, controller

+++ src/main/java/com/icia/web/dao/WDRezDao.java
@@ -27,11 +27,13 @@ public interface WDRezDao {
 	
 	public long rezFullPriceInsert(WDRez wdRez);
 	
-	public long rezStudioUpdate(WDRez wdRez);
+	//업데이트문은 건수로 리턴하는데 long타입은 너무 불필요하게 큰 데이터 타입이라 int로 변경합니다.
+	public int rezStudioUpdate(WDRez wdRez);
 	   
-	public long rezDressUpdate(WDRez wdRez);
+	public int rezDressUpdate(WDRez wdRez);
 	   
-	public long rezMakeupUpdate(WDRez wdRez);
+	public int rezMakeupUpdate(WDRez wdRez);
 	   
-	public long rezHallUpdate(WDRez wdRez);
+	public int rezHallUpdate(WDRez wdRez);
+	 
 }
